requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.3
python-dateutil==2.8.2
fpdf2==2.7.5
xhtml2pdf==0.2.4
psycopg2
psycopg